+++ src/lib.rs
@@ -267,7 +267,7 @@ impl Speaker {
         handle_error(unsafe {
             bindings::espeak_ng_Synthesize(
                 text_nul_term.as_ptr().cast::<std::ffi::c_void>(),
-                text_nul_term.len() as u64,
+                text_nul_term.len(),
                 0,
                 bindings::espeak_POSITION_TYPE_POS_CHARACTER,
                 0,
